Move `environment` package
