+++ common/persistence/persistence-tests/persistence_test_base.go
@@ -58,7 +58,7 @@ import (
 	"go.temporal.io/server/common/resolver"
 	"go.temporal.io/server/common/searchattribute"
 	"go.temporal.io/server/common/telemetry"
-	"go.temporal.io/server/environment"
+	"go.temporal.io/server/temporal/environment"
 )
 
 // TimePrecision is needed to account for database timestamp precision.
