+++ common/persistence/tests/postgresql_test_util.go
@@ -39,7 +39,7 @@ import (
 	"go.temporal.io/server/common/persistence/sql/sqlplugin"
 	"go.temporal.io/server/common/resolver"
 	"go.temporal.io/server/common/shuffle"
-	"go.temporal.io/server/environment"
+	"go.temporal.io/server/temporal/environment"
 	"go.uber.org/zap/zaptest"
 )
