+++ tests/ndc/ndc_test.go
@@ -63,8 +63,8 @@ import (
 	"go.temporal.io/server/common/persistence/versionhistory"
 	test "go.temporal.io/server/common/testing"
 	"go.temporal.io/server/common/testing/protorequire"
-	"go.temporal.io/server/environment"
 	"go.temporal.io/server/service/history/consts"
+	"go.temporal.io/server/temporal/environment"
 	"go.temporal.io/server/tests/testcore"
 	"go.uber.org/mock/gomock"
 	"google.golang.org/grpc"
