+++ tests/ndc/replication_task_batching_test.go
@@ -54,8 +54,8 @@ import (
 	"go.temporal.io/server/common/persistence/serialization"
 	test "go.temporal.io/server/common/testing"
 	"go.temporal.io/server/common/testing/protorequire"
-	"go.temporal.io/server/environment"
 	"go.temporal.io/server/service/history/replication/eventhandler"
+	"go.temporal.io/server/temporal/environment"
 	"go.temporal.io/server/tests/testcore"
 	"go.uber.org/mock/gomock"
 	"google.golang.org/protobuf/types/known/durationpb"
