+++ tools/cassandra/cqlclient_tests.go
@@ -28,7 +28,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/log/tag"
-	"go.temporal.io/server/environment"
+	"go.temporal.io/server/temporal/environment"
 	"go.temporal.io/server/tools/common/schema/test"
 )
