+++ tools/cassandra/main.go
@@ -30,8 +30,8 @@ import (
 
 	"github.com/urfave/cli"
 	"go.temporal.io/server/common/log"
-	"go.temporal.io/server/environment"
 	dbschemas "go.temporal.io/server/schema"
+	"go.temporal.io/server/temporal/environment"
 	"go.temporal.io/server/tools/common/schema"
 )
