+++ tools/cassandra/setup_task_tests.go
@@ -28,7 +28,7 @@ import (
 	"os"
 
 	"go.temporal.io/server/common/log/tag"
-	"go.temporal.io/server/environment"
+	"go.temporal.io/server/temporal/environment"
 	"go.temporal.io/server/tools/common/schema/test"
 )
