+++ tools/tests/mysql_cli_test.go
@@ -30,8 +30,8 @@ import (
 
 	"github.com/stretchr/testify/suite"
 	"go.temporal.io/server/common/persistence/sql/sqlplugin/mysql"
-	"go.temporal.io/server/environment"
 	mysqlversionV8 "go.temporal.io/server/schema/mysql/v8"
+	"go.temporal.io/server/temporal/environment"
 	"go.temporal.io/server/tools/sql/clitest"
 )
